Add support for compilation options
Add options to the Compile function, and in turn, the Run function
in order to modify the compiled command instance before executing
it. This is helpful to, for example, specify that the command
should be run in a separate process group in order for ffmpeg not
to react on SIGINTs sent to the parent proces.